backend/sub_apps/bisma/fastapi/requirements.txt
python-multipart
python-dotenv
fastapi 
uvicorn
sqlalchemy 
psycopg2-binary
pydantic
pandas
numpy 
openpyxl